+++ src/components/pages/FAQ/index.jsx
@@ -1,5 +1,5 @@
-import UserMenu from '../UI/organisms/UserMenu'
-import Sidebar from '../UI/organisms/Sidebar'
+import UserMenu from '../../UI/organisms/UserMenu'
+import Sidebar from '../../UI/organisms/Sidebar'
 
 import { UnverifiedIcon } from '@primer/octicons-react'
